remove arm64 from tests, simplify read
