+++ .github/workflows/ci.yml
@@ -26,7 +26,8 @@ jobs:
   test:
     strategy:
       matrix:
-        os: [ubuntu-22.04, macos-14, windows-2022, ubuntu-22.04-arm64]
+        #os: [ubuntu-22.04, macos-14, windows-2022, ubuntu-22.04-arm64]
+        os: [ubuntu-22.04, macos-14, windows-2022]
     runs-on: ${{ matrix.os }}
     steps:
       - name: Install build-essential and clang (arm64)
